requirements.txt
blinker==1.6.2
boto3==1.28.46
botocore==1.31.46
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
gunicorn==20.1.0
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
python-dotenv==0.19.1
openai
numpy
tiktoken
pgvector
psycopg
psycopg-binary